Vendor `gardener@1.81.7`, Upgrade dependencies and implement `ForceDelete` method for extensions (#297)
* Vendor gardener@v1.81.7

* Implement ForceDelete in controllers

* Adapt breaking changes in CP

* Replace semver with semver/v3

+++ pkg/controller/controlplane/add.go
@@ -60,7 +60,7 @@ func AddToManagerWithOptions(ctx context.Context, mgr manager.Manager, opts AddO
 		nil, shootAccessSecretsFunc, nil, nil,
 		nil, controlPlaneChart, controlPlaneShootChart, nil, storageClassChart, nil,
 		NewValuesProvider(mgr), extensionscontroller.ChartRendererFactoryFunc(util.NewChartRendererForShoot),
-		imagevector.ImageVector(), "", opts.ShootWebhookConfig, opts.WebhookServerNamespace, defaultServer.Options.Port)
+		imagevector.ImageVector(), "", opts.ShootWebhookConfig, opts.WebhookServerNamespace, int32(defaultServer.Options.Port))
 	if err != nil {
 		return err
 	}
